requirements.txt
python-dotenv~=0.18.0
requests~=2.26.0
pysimplegui~=4.47.0